+++ src/utils/swaggerUtil.ts
@@ -199,8 +199,6 @@ const collectV2AllDepDefs = (defQueue: DepDefListItem[], entireDefs: OpenAPIV2.D
 export const shakeV2RefsInSchema = (schema: OpenAPIV2.SchemaObject, entireDefs: OpenAPIV2.DefinitionsObject) => {
   // 先收集 schema 中的所有 $ref 实体名称
   const schemaRefNameSet = collectRefName({ ...schema, definitions: null });
-  console.log('******schemaRefNameSet', schemaRefNameSet);
-  console.log('******entireDefs', entireDefs);
   // 再根据从 schema 中收集到的实体类对 definitions 进行二次收集
   const defNameSet = new Set<string>();
   schemaRefNameSet.forEach((ref) => {